Refactor parser and data models
This refactors both the parser and data model classes for SPHEREx
pipeline metadata documents into base classes that can apply to all
SPHEREx documents. This will enable us to add support for project
management (PM) documents, which have slightly different (mostly a
subset) metadata of what pipeline module documents have.
