+++ src/spherexlander/parsers/pipelinemodule/__init__.py
@@ -1,6 +1,4 @@
-"""Parser for SPHEREx pipeline modoule documents."""
-
-__all__ = ["SpherexPipelineModuleParser", "SpherexPipelineModuleMetadata"]
+"""Parser for SPHEREx Module Specification (SSDC-MS) documents."""
 
 from spherexlander.parsers.pipelinemodule.datamodel import (
     SpherexPipelineModuleMetadata,
@@ -8,3 +6,5 @@ from spherexlander.parsers.pipelinemodule.datamodel import (
 from spherexlander.parsers.pipelinemodule.parser import (
     SpherexPipelineModuleParser,
 )
+
+__all__ = ["SpherexPipelineModuleParser", "SpherexPipelineModuleMetadata"]
